Add initial Open Graph product recognition.
This looks for meta tags with a 'property' attribute with a value of 'og:*' to extract product title, image and price from a page by loading a content script, productInfo.js, into every page. This patch does not distinguish a product page from other kinds of web pages. The results are logged in the browser console by background.js.