Improve data extraction method using a static ruleset. Fixes #21.
This builds off of #6, extending product recognition to also look for product information (title, image and price only so far) in a static dataset, 'product_extraction_data.json'. The fallback is to check Open Graph meta tags.

Note: 'product_extraction_data.json is a manually generated set of CSS selectors for a product page by vendor. As a proof-of-concept, I have only added four of the biggest e-commerce sites to populate this dataset.