WIP - Display some product info in the sidebar. Fixes #13.
This builds off of #6, extending product recognition to also look for product information (title, image and price only so far) in a static dataset, 'product_recognition_data.json'. The fallback is to check Open Graph meta tags. The information is logged to the browser console.

Note: 'product_recognition_data.json is a manually generated set of CSS selectors for a product page by vendor. I will just be looking at 5 of the biggest e-commerce sites around to populate this dataset.

TODO:
* Differentiate product pages from other pages to avoid unnecessary processing and unrelated data extraction (e.g. Open Graph 'titles' are often found on non-product pages)
* Send product information from background or content script to the sidebar page script to be displayed