Incorporate feedback from osmose.
Rebased on top of latest master.

Changed the structure of 'product_extraction_data.json' to add a third level key, 'extractUsing', so that we can know for a given product attribute (e.g. title) how to pull that information from an element for a particular vendor.

For example: For a Walmart product page, to obtain the product title, we want to grab an element using one of these selectors and then extract the title using the innerText value of that element.