Minor tweaks per osmose's additional feedback.
Checks for tab.status === 'loading' in addition to the existence of 'changeInfo.url' for the 'tabs.onUpdated' callback to debounce the location change event a little more. Also made SIDEBAR_URL a global variable, so it can be used both when the sidebar page is reset on location change and when we are setting a new sidebar URL for a product page.