Fix #28: Add Firefox/Marionette/Tape-based test suite.
Adds an automated test suite that runs JS tests using Tape. The tests are
run in Firefox using the remote code execution capabilities of Marionette.

Marionette's only well-maintained client is in Python, so we also have to add
Python dependencies, including a Pipfile.

bin/run_tests.py contains most of the plumbing for getting the test JS running,
including bundling it with Webpack, launching Firefox, and formatting the
output.