Add score threshold for Fathom extraction.
Since it's possible for multiple elements to be returned by Fathom for a given ruleset applied to a page, I applied a threshold (currently just set as the score for the only rule we have, 'hasDivWithPriceClass') to help provide an additional measure of confidence.

If more than one element meet or exceed the threshold, the first element in the list is returned.