Incorporate feedback from Osmose
I got the closing parenthesis to show up, though not with the ::after pseudoelement. This was the only way I could get it to work. Looks right though!

I gave the .undo-label-wrapper a 'display: flex' so that its children would be flex items whose size would adjust. Its children are .undo-label and .title-wrapper. Since .title-wrapper overflows its parent container .undo-label-wrapper, we see the ellipsis, and since the closing parenthesis is not a flex item (or part of the .title-wrapper or .title), it shows up without being cut off itself.