Fix #112, fix #60: Dismiss price alerts when browser action closes.
Switch to using a messaging port disconnect to detect when the browser action
closes. Without this, dismissing price alerts is not consistent as the JS
context for the browser action is unloaded before the async work of updating
the store is finished.

Because we are no longer performing async work after the browser action unloads,
we no longer see the errors from #60 about promises resolving after the context
is destroyed.