Call updatePrices recursively with setTimeout
[Reference](https://developer.mozilla.org/en-US/docs/Web/API/WindowOrworkerGlobalScope/setInterval#Ensure_that_execution_duration_is_shorter_than_interval_frequency)

`updatePrices` is now only called once in the main background script and calls itself recursively to ensure the previous call to `updatePrices` has finished before the next one is called. This avoids the possibility that calls to updatePrices will overlap at the cost of non-regular checks (i.e. the interval between `updatePrices` calls is no longer simply `priceCheckTimeoutInterval`, but `priceCheckTimeoutInterval` + however long it takes to execute `updatePrices`, which is roughly `n * iframeTimeout`, where `n` is the number of tracked products).