Catch connection error for extracted product popup updates
If the popup is closed when extraction completes on a page, the background script tries to send a message to the popup to update it, which results in an error:
```
Error: Could not establish connection. Receiving end does not exist.
```

This patch simply catches that error, so that we don't see it in the browser console.