Incorporate Osmose's feedback
Replaced 'isValidResource' indirection with easier-to-follow validation inline.

'url' and 'imageUrl' are declared outside of the 'try...catch' statement, so that the only error we catch in the catch block is from an error in creating the URL instance.

Renamed 'DATA_URI_PATHS' to 'ALLOWED_IMAGE_MIMETYPES' to be more descriptive of the value and add some helpful comments.