Fix JavaScript error introduced by PR #247
Tacking this on, though it's an unrelated error.

Previously, we were attempting to send messages via `tabs.sendMessage` to all content scripts both the ones running in content processes and in the background. However, the tabId for a content script in the background is -1, which is not a valid value for sending messages to content scripts in specific tabs.

Since the code that introduced this error was intended to fix an issue related to badging the browserAction toolbar button to add a new product, and since that badging is only triggered by extraction in content processes, we don't need to worry about sending the message to the background.

Now, we now no longer see the error:
`Type error for parameter tabId (Integer -1 is too small (must be at least 0)) for tabs.sendMessage.`