Fix #177: Disable price polling in Private Browsing windows
Update `shouldUpdatePrices` with a check for whether or not the active window is a private window.

I did not add a check for `isActiveWindowPrivate` in `shouldExtract` in `./src/privacy.js`, despite the fact that `browser.extension.inIncognitoContext` returns the "wrong" value for background extraction, because background extraction is gated by the `shouldUpdatePrices` check.