Adds fix for image candidates that have 1-D bounding boxes
Sometimes a feature vector can output a 'null' value. This will [throw an error](https://github.com/mozilla/fathom-fox/issues/35) during training with the new 'fathom-train' CLI.

Possible causes include:
* A name mismatch between the 'name' value passed into a rule function (its second argument) and the name of the rule in the list of [ruleName, coefficient] tuples referenced in the ruleset object.
* A score callback might be failing to return a number.
* A corner case of a DOM or CSSOM specification used by a score callback; e.g. innerText could return null instead of the empty string in Firefox.

In this case, a score callback was failing to return a number if the width or height of the element passed into 'aspectRatio' was 0.