Fixes rule name mismatch for price regex.
As mentioned in [this issue](https://github.com/mozilla/fathom-fox/issues/35), it's possible that a feature vector will contain a 'null' value for one or more rules, which will cause 'fathom-train' to throw an error.

In this case, the reason for the 'null' value was a name mismatch.

I opted to change the method name as its naming convention did not match the other rules.