Replaces vectors depending on image coefficients and bias
In Price Tracker, the product 'title' and 'price' features are dependent upon the 'image' feature results (e.g. there is a rule for the 'price' feature called 'isNearImage' which scores a candidate 'price' element based on its proximity to the most likely 'image' element).

As a result, the final weights and bias from training the 'image' feature need to be taken into account before vectorizing the 'title' and 'price' features. This commit updates the vectors for 'title' and 'price' to take this into account.